+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-（なし）
+- **Breaking**: 最小サポートをbeta53に引き上げ
+- **Breaking**: module: パラメーター取得系の関数を`Option`から`Result<T, GetParamError>`にする
+  - Note: `#[aviutl2::module::functions]`を使っていて、かつ`#[direct]`を使用していない場合は追従は不要です。
 
 ### デモプラグイン